+++ docs/en/changelog.rst
@@ -25,7 +25,9 @@ Core
 
 * Added velocities enumerations to IkParameterizationType (like Transform6DVelocity), coordinate transforms can handle velocities.
 
-* :meth:`.IkSolver.Solve` can output custom data returned from IK filters as a :class:`.IkFilterReturn` object.
+* :meth:`.IkSolver.Solve` can output custom data returned from IK filters as a :class:`.IkReturn` object.
+
+* Added IK solver error codes for notifying user where IK failed.
 
 Database
 --------